# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for UniMatch Copilot AI system",
        "title": "UniMatch Copilot API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/cases": {
            "get": {
                "description": "Fetch paginated list of cases with optional status filter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases"
                ],
                "summary": "List cases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by case status: pending, processing, done, human_review, failed",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "meta": {
                                            "$ref": "#/definitions/response.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create student and case records, triggers AI profiling",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases"
                ],
                "summary": "Create a new case",
                "parameters": [
                    {
                        "description": "Student Information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCaseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CaseCreatedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/cases/count": {
            "get": {
                "description": "Get total cases count, optionally filtered by status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases"
                ],
                "summary": "Count cases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by case status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/cases/{id}": {
            "get": {
                "description": "Retrieve a case and its active recommendations by UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases"
                ],
                "summary": "Get case details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Case UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/cases/{id}/report": {
            "post": {
                "description": "Queue an async report generation job for a ready case",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases"
                ],
                "summary": "Request PDF report generation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Case UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ReportStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CaseCreatedResponse": {
            "type": "object",
            "properties": {
                "case_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateCaseRequest": {
            "type": "object",
            "required": [
                "budget_usd_per_year",
                "full_name",
                "gpa_normalized"
            ],
            "properties": {
                "achievements": {
                    "type": "string"
                },
                "budget_usd_per_year": {
                    "type": "integer",
                    "minimum": 0
                },
                "extracurriculars": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gpa_normalized": {
                    "type": "number",
                    "maximum": 4,
                    "minimum": 0
                },
                "gpa_raw": {
                    "type": "number"
                },
                "gpa_scale": {
                    "type": "number"
                },
                "ielts_breakdown": {
                    "$ref": "#/definitions/dto.IeltsBreakdown"
                },
                "ielts_overall": {
                    "type": "number"
                },
                "intended_major": {
                    "type": "string"
                },
                "personal_statement_notes": {
                    "type": "string"
                },
                "preferred_countries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sat_total": {
                    "type": "integer"
                },
                "scholarship_required": {
                    "type": "boolean"
                },
                "target_intake": {
                    "type": "string"
                },
                "toefl_total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.IeltsBreakdown": {
            "type": "object",
            "properties": {
                "listening": {
                    "type": "number"
                },
                "reading": {
                    "type": "number"
                },
                "speaking": {
                    "type": "number"
                },
                "writing": {
                    "type": "number"
                }
            }
        },
        "dto.ReportStatusResponse": {
            "type": "object",
            "properties": {
                "case_id": {
                    "type": "string"
                },
                "report_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean"
                },
                "has_prev": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorInfo"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "swagger.SwaggerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "details": {
                    "type": "string",
                    "example": "User with ID 123 not found"
                },
                "message": {
                    "type": "string",
                    "example": "Resource not found"
                }
            }
        }
    }
}